fixed the code duplication as per sonarqube
